Updating tests : test_main
